finished part-38 (designing the "login" page)
Here we have designed the "users_area/user_login.php" file at first. Then moved the "checkout.php" to "users_area/checkout.php" so that we can display the login page layout inside the checkout page; without redirecting to a separate login page (only for the "checkout" page). After that, we linked the "login" and "registration" page to the remaing files:
- index.php
- display_all.php
- product_details.php
- cart.php
so that, we can access to the login and registration page from them.