finished part-40 (conditions for "name", "email", and "passwords")
Here in the "users_area/user_registration.php" file we give some conditions to check whether the "username" or "email" is already stored in the database or not. And we also checked whether the "password" and "confirmed password" matches or not. The registered information is only stored in the database table "user_table" when both the conditions are checked.