finished part-43 (checking user login credentials)
Here in the "users_area/user_login.php" file we make php changes to check the user credentials. We first check whether the "username" followed by the "password" given now to login, matches with the data stored in our database during the user registration. When both the "username" and "password" matches, onnly then the user is able to "successfully login".