finished part-49 (creating "user_orders" and "orders_pending" tables in our database)
Here we have created two new tables. One is the "orders_pending" with-
- order_id
- user_id
- invoice_number
- product_id
- quantity
- order_status
where if the user has orders in the cart and completed payment then status is "completed", otherwise "pending". The next table is "user_orders" with-
- order_id
- user_id
- amount_due
- invoice number
- total products
- order_date
- order_status
where, whenever the user clicks on the "pay offline" button; all the items in the cart of that user, will be added to the database.