finished part-52 (removing items from cart, after placing order)
We have updated the "users_area/order.php" file, so that when placing an order; the items from the cart will be removed automatically. And also, during order placing; the data will be added to the "user_orders" and "orders_pending" tables simultaniously. But for now, the problem here is; only one product is added to the "orders_pending" table in the database--will work on it later.