finished part-59 (designing the "Online Payment" page to pay)
Here we have created a new file "users_area/confirm_payment.php" for the user to select their payment option. In this page, the order "invoice_number" and "price" will automatically show in the form. There is also various online options to select from for the user.
Additionally, in this part, we also modified the "index.php" file to show the "Account" info if the user is logged in; and "Registration" info when is not.